Rename tests to create
