+++ SummerBornInfo.Web.Test/Features/EstablishmentStatusCommands/EstablishmentStatus_Create.cs
@@ -1,27 +1,27 @@
-﻿using Create = SummerBornInfo.Web.Features.PhaseOfEducationCommands.Create;
+﻿using Create = SummerBornInfo.Web.Features.EstablishmentStatusCommands.Create;
 
-namespace SummerBornInfo.Web.Test.Features.PhaseOfEducationCommands;
-public class PhaseOfEducation_Commands(PostgresTestFixture App) : BaseIntegrationTest
+namespace SummerBornInfo.Web.Test.Features.EstablishmentStatusCommands;
+public class EstablishmentStatus_Create(PostgresTestFixture App) : BaseIntegrationTest
 {
     [Fact]
-    public async Task ValidRequest_Create_SavesAuthoritySuccessfully()
+    public async Task ValidRequest_Create_SavesStatusSuccessfully()
     {
         var (rsp, res) = await App.Client.POSTAsync<Create.Endpoint, Create.Request, Create.Response>(new()
         {
-            Code = 500,
+            Code = 200,
             Name = "Test"
         });
         rsp.StatusCode.Should().Be(HttpStatusCode.OK);
         res.Id.Should().NotBeEmpty();
-        res.Code.Should().Be(500);
+        res.Code.Should().Be(200);
         res.Name.Should().Be("Test");
 
         using var scope = App.Services.CreateScope();
         await using var dbContext = scope.ServiceProvider.GetRequiredService<SchoolContext>();
-        var savedGroup = await dbContext.PhaseOfEducation.AsNoTracking().SingleAsync(g => g.Id == res.Id);
+        var savedGroup = await dbContext.EstablishmentStatus.AsNoTracking().SingleAsync(g => g.Id == res.Id);
         savedGroup.Should().NotBeNull();
         savedGroup.Id.Should().Be(res.Id);
-        savedGroup.Code.Should().Be(500);
+        savedGroup.Code.Should().Be(200);
         savedGroup.Name.Should().Be("Test");
     }
